+++ CHANGELOG.md
@@ -44,6 +44,8 @@ New features:
 
 Bug fixes:
 * Fix panic when a file name starts with a multibyte UTF-8 character
+* Update the `..` entry when moving a directory to a different parent
+
 * Fix formatting volumes with size in range 4096-4199 KB
 * Always respect `fat_type` from `FormatVolumeOptions`
 * Fill FAT32 root directory clusters with zeros after allocation to avoid interpreting old data as directory entries